refactor: move Legacy provider to provider/legacy sub-package
