+++ cmd/providers_register.go
@@ -17,4 +17,5 @@ package cmd
 // Blank imports to trigger init() registration for provider sub-packages.
 import (
 	_ "github.com/penny-vault/pvdata/provider/fred"
+	_ "github.com/penny-vault/pvdata/provider/legacy"
 )
